+++ src/RentingManager.cpp
@@ -22,7 +22,7 @@ RentingManager::~RentingManager() {
 
 }
 
-double RentingManager::getSize() {
+int RentingManager::getSize() {
     return RentingManager::rentedItems.size();
 }
 
@@ -30,6 +30,6 @@ bItem RentingManager::getRent(int id) {
     return rentedItems.at(id);
 }
 
-double RentingManager::getIDRent() {
+int RentingManager::getIDRent() {
     return RentingManager::idRent;
 }
